+++ CHANGELOG.md
@@ -45,7 +45,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `Zedex.replace/1` and `Zedex.replace_with/2`
   - Replace entire functions in a module
-- `Zedex.replace_calls/3`
+- `Zedex.Danger.replace_calls/3`
   - Replace calls in a function with calls to another function
 - `Zedex.reset/0` and `Zedex.reset/1`
   - Reset patched modules back to their original state